requirements.txt
langchain==0.3.25
langchain-openai==0.3.12
langchain-community==0.3.24
chromadb==1.0.7
tiktoken==0.9.0
pypdf==5.6.0
python-dotenv==1.1.0
faiss-cpu==1.11.0
